+++ apgi_framework/analysis/statistical_report_generator.py
@@ -532,7 +532,7 @@ class StatisticalReportGenerator:
         except (ValueError, RuntimeError, AttributeError) as e:
             # Fallback approximation
             self.logger.warning(f"Power calculation failed: {e}. Using approximation.")
-            return min(1.0, (abs(effect_size) * np.sqrt(sample_size)) / 2.8)
+            return float(min(1.0, (abs(effect_size) * np.sqrt(sample_size)) / 2.8))
 
     def _generate_overall_conclusion(
         self,
